Scroll bug between pages, new window icon for resume
When switching between longer pages (Photography/Work), being scrolled down the page would maintain where you scrolled. Now it resets users to the top of the window when the component is mounted.

Added an icon to the navigation bar to show that my resume opens in a new window. Might consider doing this for all outgoing links, but that could get cluttered.

Fixed About page (again) for presentation, didn't like how it was before. Also renamed the image to be less specific.

className is now actually passed down to the `Fade` element when provided to a `PageTemplate`. That didn't happen before.